Deleting user additions. 1. Now deleting user also deleting his categories and operations.

+++ src/server/routes/user.ts
@@ -6,6 +6,9 @@ import serverError from '../../utils/errorsToClient/serverError.ts';
 import { redLog } from '../../utils/console/coloredLogs.ts';
 import showError from '../../utils/console/showError.ts';
 import sendAuthError from '../../utils/errorsToClient/sendAuthError.ts';
+import { Mongoose, Types } from 'mongoose';
+import Operation from '../models/Operation.ts';
+import Category from '../models/Category.ts';
 
 const router = express.Router({ mergeParams: true });
 
@@ -36,18 +39,23 @@ async function removeUser(req: AuthedRequest, res: Response) {
     if (!removingUser) { { sendNotFoundError(res, id); return; }; }
     const isPermitted = (removingUser._id.toString() === req.user._id);
     if (!isPermitted) return sendAuthError(res, 'user/remove', req.user._id);
-    await User.findByIdAndDelete(removingUser._id);
-    await tokenService.removeTokens(removingUser._id);
-    // await removeOperations(removingUser._id); // TODO turn on when ready
-    // await removeCategories(removingUser._id);
-    res.status(200).send(null);
+
+    await removeUserData(removingUser._id, Operation);
+    await removeUserData(removingUser._id, Category);
+    const result = await removingUser.deleteOne();
+    if(result.deletedCount === 1) await tokenService.removeTokens(removingUser._id);
+
+    res.status(200).send(result);
   } catch (e) {
     showError(e);
     serverError(res, 'auth/user/removeUser');
   }
 }
 
-
+async function removeUserData(id: Types.ObjectId, model: Mongoose["Model"]) {
+  const userData = await model.find({ user: id });
+  return await Promise.all(userData.map((document) => document.deleteOne()));
+}
 function sendNotFoundError(res: Response, id: string) {
   res.status(404).json({ message: `User to rmove not found` });
   redLog(`User with id ${id} not found`);
